+++ tests/regression/rm3/test_rm3_reg_waves.cpp
@@ -206,7 +206,7 @@ int main(int argc, char* argv[]) {
             outputFile << std::left << std::setw(10) << "Time (s)" << std::right << std::setw(16) << "Float Heave (m)"
                        << std::right << std::setw(16) << "Plate Heave (m)" << std::right << std::setw(16)
                        << "Float Drift (x) (m)" << std::endl;
-            for (int i = 0; i < time_vector.size(); ++i)
+            for (size_t i = 0; i < time_vector.size(); ++i)
                 outputFile << std::left << std::setw(10) << std::setprecision(2) << std::fixed << time_vector[i]
                            << std::right << std::setw(16) << std::setprecision(4) << std::fixed
                            << float_heave_position[i] << std::right << std::setw(16) << std::setprecision(4)
