+++ tests/regression/sphere/test_sphere_decay.cpp
@@ -138,7 +138,7 @@ int main(int argc, char* argv[]) {
                        //<< std::right << std::setw(18) << "Heave Vel (m/s)"
                        //<< std::right << std::setw(18) << "Heave Force (N)"
                        << std::endl;
-            for (int i = 0; i < time_vector.size(); ++i)
+            for (size_t i = 0; i < time_vector.size(); ++i)
                 outputFile << std::left << std::setw(12) << std::setprecision(6) << std::fixed << time_vector[i]
                            << std::right << std::setw(12) << std::setprecision(6) << std::fixed << heave_position[i]
                            << std::endl;
